move opa files into their own folder
